# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THREE NINE BAR Inventory System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /wasting:
    get:
      operationId: listWasting
      tags: [wasting]
      summary: List wasting records
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of wasting records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wasting"
    post:
      operationId: createWasting
      tags: [wasting]
      summary: Record wasting (auto-deducts stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WastingInput"
      responses:
        "201":
          description: Wasting recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wasting"
  /wasting/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateWasting
      tags: [wasting]
      summary: Update wasting record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WastingInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wasting"
    delete:
      operationId: deleteWasting
      tags: [wasting]
      summary: Delete wasting record (restores stock)
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /stock/movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: List stock movements
      parameters:
        - name: ingredientId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
  /stock/add:
    post:
      operationId: addStock
      tags: [stock]
      summary: Add stock (IN movement)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAddInput"
      responses:
        "201":
          description: Stock added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      required: [id, username, role, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, staff]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff]
        isActive:
          type: boolean
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff]
        isActive:
          type: boolean
...
    Wasting:
      type: object
      required: [id, wastingDate, ingredientId, ingredientName, qty, unit, reason]
      properties:
        id:
          type: integer
        wastingDate:
          type: string
        ingredientId:
          type: integer
        ingredientName:
          type: string
        qty:
          type: number
        unit:
          type: string
        reason:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    WastingInput:
      type: object
      required: [wastingDate, ingredientId, qty, reason]
      properties:
        wastingDate:
          type: string
        ingredientId:
          type: integer
        qty:
          type: number
        reason:
          type: string
        notes:
          type: string
...
    StockMovement:
      type: object
      required: [id, ingredientId, ingredientName, movementType, qty, unit, referenceType, createdAt]
      properties:
        id:
          type: integer
        ingredientId:
          type: integer
        ingredientName:
          type: string
        movementType:
          type: string
          enum: [in, out, wasting]
        qty:
          type: number
        unit:
          type: string
        referenceType:
          type: string
        referenceId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    StockAddInput:
      type: object
      required: [ingredientId, qty]
      properties:
        ingredientId:
          type: integer
        qty:
          type: number
        notes:
          type: string
...