# lib/api-zod/src/generated/types/stockAddInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE BAR Inventory System API
 * OpenAPI spec version: 0.1.0
 */

export interface StockAddInput {
  ingredientId: number;
  qty: number;
  notes?: string;
}