# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE BAR Inventory System API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementMovementType } from './stockMovementMovementType';

export interface StockMovement {
  id: number;
  ingredientId: number;
  ingredientName: string;
  movementType: StockMovementMovementType;
  qty: number;
  unit: string;
  referenceType: string;
  /** @nullable */
  referenceId?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}